@@ -1,4 +1,4 @@
-package com.tangentem.myapplication
+package com.example.myapplication
 
 import org.junit.Test
 
